+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # [1.2.3-rc.2]
 
-- Fixed RainViewer weather not loading on some zoom levels, improved its loading speed
+- Fixed RainViewer weather not loading on some zoom levels
 
 # [1.2.3-rc.1]